requirements.txt
# requirements.txt
Pillow
exifread